==> 2_Data_wrangling/10_missings.qmd <==
---
title: "Missing values"
author: "Solveig Bjørkholt"
---

## The dataset we are working on (after data cleaning in previous chapter):

```{r, echo=FALSE, message = FALSE, warning=FALSE, error = FALSE}
library(tidyverse)
load("../datafolder/costofliving2.rda")
costofliving <- as_tibble(costofliving)
DT::datatable(costofliving)
```

```{r}
costofliving %>%
  head()
```

What are those `NA` things that occur in some cells in the dataset? They are called "missing values". Missing values occur when there are values in the dataset that should ideally have been there, but they're not. This could for example be because we didn't find data on that particular thing, somebody refused to give us the information we needed, or because we've done something wrong in the code, generating `NA`.

To see how much missing there is in our dataset, we can use the function `is.na` along with the function `table`. `is.na` gives the value `TRUE` if the cell has a missing value and `FALSE` otherwise, while the function `table` counts the instances in which `TRUE` and `FALSE` occurs. This means, in other words, that we have 682 missing values in our data.

```{r}
costofliving %>%
  is.na() %>%
  table()
```

Want to remove all rows that have one or more missing values? Use the function `na.omit`.

```{r}
costofliving %>%
  na.omit()
```

If you would like to more carefully select which variables you'd like to remove the missing values from, use `drop_na` and choose the variables where the rows with missing values should be removed.

```{r}
costofliving %>%
  drop_na(country, city, cofi)
```

```{r, echo = FALSE}
costofliving2 <- costofliving %>%
  drop_na %>%
  mutate(city = str_squish(city),
         country = str_squish(country))

saveRDS(costofliving2, file = "../datafolder/costofliving2.rds")
```
